move fetchBlockFagDecoreated to utils/loadBlock
