+++ express/scripts/utils/decorate.js
@@ -1,7 +1,4 @@
-import { getLibs } from '../utils.js';
 import { getIconElement } from './icons.js';
-
-const { createTag, getConfig } = await import(`${getLibs()}/utils/utils.js`);
 
 // This was only added for the blocks premigration. It is not to be used for new blocks.
 export function decorateButtonsDeprecated(el) {
@@ -83,32 +80,6 @@ export function normalizeHeadings(block, allowedHeadings) {
       }
     }
   });
-}
-
-export async function fetchBlockFragDecorated(url, blockName) {
-  const location = new URL(window.location);
-  const { prefix } = getConfig().locale;
-  const fragmentUrl = `${location.origin}${prefix}${url}`;
-
-  const path = new URL(fragmentUrl).pathname.split('.')[0];
-  const resp = await fetch(`${path}.plain.html`);
-  if (resp.status === 404) {
-    return null;
-  }
-
-  const html = await resp.text();
-  const section = createTag('div');
-  section.innerHTML = html;
-  section.className = `section section-wrapper ${blockName}-container`;
-  const block = section.querySelector(`.${blockName}`);
-  block.dataset.blockName = blockName;
-  block.parentElement.className = `${blockName}-wrapper`;
-  block.classList.add('block');
-  const img = section.querySelector('img');
-  if (img) {
-    img.setAttribute('loading', 'lazy');
-  }
-  return section;
 }
 
 export function decorateSocialIcons($main) {
